+++ src/Shiny.Core/Platforms/Android/AndroidShinyHost.cs
@@ -11,9 +11,6 @@ using Microsoft.Extensions.DependencyInjection.Extensions;
 using Android.Runtime;
 using Android.Content;
 
-[assembly: UsesPermission(Android.Manifest.Permission.AccessNetworkState)]
-[assembly: UsesPermission(Android.Manifest.Permission.BatteryStats)]
-[assembly: UsesPermission(Android.Manifest.Permission.ReceiveBootCompleted)]
 
 namespace Shiny
 {