+++ src/Shiny.Net.Http/LogCategory.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+
+namespace Shiny.Net.Http
+{
+    public static class LogCategory
+    {
+    }
+}