+++ src/Shiny.Notifications/Platforms/Android/NotificationManager.cs
@@ -6,23 +6,14 @@ using System.Threading.Tasks;
 using Android.App;
 using Android.Content;
 using Android.Media;
-using Android.Support.V4.App;
+using Android.Graphics;
+using AndroidX.Core.App;
+using AndroidX.Core.Content;
 using Shiny.Infrastructure;
 using Shiny.Jobs;
 using Shiny.Logging;
 using Shiny.Settings;
-using Native = Android.App.NotificationManager;
-using Android.Graphics;
 
-#if ANDROIDX
-using AndroidX.Core.App;
-using RemoteInput = AndroidX.Core.App.RemoteInput;
-using TaskStackBuilder = AndroidX.Core.App.TaskStackBuilder;
-#else
-using Android.Support.V4.App;
-using RemoteInput = Android.Support.V4.App.RemoteInput;
-using TaskStackBuilder = Android.App.TaskStackBuilder;
-#endif
 
 namespace Shiny.Notifications
 {
@@ -34,9 +25,7 @@ namespace Shiny.Notifications
         readonly ISettings settings;
         readonly ISerializer serializer;
         readonly IJobManager jobs;
-
         readonly NotificationManagerCompat? compatManager;
-        readonly Native? newManager;
 
 
         public NotificationManager(AndroidContext context,
@@ -53,14 +42,7 @@ namespace Shiny.Notifications
             this.repository = repository;
             this.settings = settings;
 
-#if ANDROIDX
             this.compatManager = NotificationManagerCompat.From(this.context.AppContext);
-#else
-            if (this.context.IsMinApiLevel(26))            
-                this.newManager = Native.FromContext(this.context.AppContext);
-            else
-                this.compatManager = NotificationManagerCompat.From(this.context.AppContext);
-#endif
             this.context
                 .WhenIntentReceived()
                 .Subscribe(x => this
@@ -79,16 +61,14 @@ namespace Shiny.Notifications
 
         public async Task Cancel(int id)
         {
-            this.newManager?.Cancel(id);
-            this.compatManager?.Cancel(id);
+            this.compatManager.Cancel(id);
             await this.repository.Remove<Notification>(id.ToString());
         }
 
 
         public async Task Clear()
         {
-            this.newManager?.CancelAll();
-            this.compatManager?.CancelAll();
+            this.compatManager.CancelAll();
             await this.repository.Clear<Notification>();
         }
 
@@ -99,17 +79,13 @@ namespace Shiny.Notifications
 
         public async Task<AccessState> RequestAccess()
         {
-            if (!this.compatManager?.AreNotificationsEnabled() ?? false)
-                return AccessState.Disabled;
-
-            if (!this.newManager?.AreNotificationsEnabled() ?? false)
+            if (this.compatManager.AreNotificationsEnabled())
                 return AccessState.Disabled;
 
             return await this.jobs.RequestAccess();
         }
 
 
-        //https://stackoverflow.com/questions/45462666/notificationcompat-builder-deprecated-in-android-o
         public async Task Send(Notification notification)
         {
             if (notification.Id == 0)
@@ -121,13 +97,36 @@ namespace Shiny.Notifications
                 return;
             }
 
+            var native = this.CreateNativeNotification(notification);
+            this.compatManager.Notify(notification.Id, native);
+            await this.services.SafeResolveAndExecute<INotificationDelegate>(x => x.OnReceived(notification), false);
+        }
+
+
+        public int Badge { get; set; }
+
+        readonly List<NotificationCategory> registeredCategories = new List<NotificationCategory>();
+        public void RegisterCategory(NotificationCategory category) => this.registeredCategories.Add(category);
+
+
+        public virtual Android.App.Notification CreateNativeNotification(Notification notification)
+        {
             var pendingIntent = this.GetLaunchPendingIntent(notification);
             var builder = new NotificationCompat.Builder(this.context.AppContext)
-                .SetContentTitle(notification.Title)                
+                .SetContentTitle(notification.Title)
                 .SetSmallIcon(this.GetSmallIconResource(notification))
                 .SetAutoCancel(notification.Android.AutoCancel)
                 .SetOngoing(notification.Android.OnGoing)
                 .SetContentIntent(pendingIntent);
+
+            if (!notification.Android.ContentInfo.IsEmpty())
+                builder.SetContentInfo(notification.Android.ContentInfo);
+
+            if (!notification.Android.Ticker.IsEmpty())
+                builder.SetTicker(notification.Android.Ticker);
+
+            if (!notification.Android.Category.IsEmpty())
+                builder.SetCategory(notification.Category);
 
             if (notification.Android.UseBigTextStyle)
                 builder.SetStyle(new NotificationCompat.BigTextStyle().BigText(notification.Message));
@@ -137,15 +136,17 @@ namespace Shiny.Notifications
             this.TrySetSound(notification, builder);
             this.TrySetLargeIconResource(notification, builder);
 
-            if (!notification.Category.IsEmpty())            
-                this.AddCategory(builder, notification);           
+            if (!notification.Category.IsEmpty())
+                this.AddCategory(builder, notification);
 
             if (notification.BadgeCount != null)
                 builder.SetNumber(notification.BadgeCount.Value);
 
+
             // disabled until System.Drawing reliable works in Xamarin again
             //if (notification.Android.Color != null)
             //    builder.SetColor(notification.Android.Color.Value)
+
             if (!notification.Android.ColorResourceName.IsEmpty())
             {
                 if (this.context.IsMinApiLevel(21))
@@ -174,20 +175,15 @@ namespace Shiny.Notifications
             if (notification.Android.Vibrate)
                 builder.SetVibrate(new long[] { 500, 500 });
 
-            this.DoNotify(builder, notification);
-            await this.services.SafeResolveAndExecute<INotificationDelegate>(x => x.OnReceived(notification), false);
+            this.CreateChannel(notification);
+            builder.SetChannelId(notification.Android.ChannelId);
+
+            return builder.Build();
         }
 
 
-        public int Badge { get; set; }
-
-        readonly List<NotificationCategory> registeredCategories = new List<NotificationCategory>();
-        public void RegisterCategory(NotificationCategory category) => this.registeredCategories.Add(category);
-
-
-        protected virtual void DoNotify(NotificationCompat.Builder builder, Notification notification)
+        protected virtual void CreateChannel(Notification notification)
         {
-#if ANDROIDX
             var channelId = notification.Android.ChannelId;
             var channel = this.compatManager.GetNotificationChannel(channelId);
 
@@ -216,51 +212,7 @@ namespace Shiny.Notifications
                 channel.EnableVibration(notification.Android.Vibrate);
                 this.compatManager.CreateNotificationChannel(channel);
             }
-            
-
-            builder.SetChannelId(channelId);
-            this.compatManager.Notify(notification.Id, builder.Build());
-#else
-            if (this.newManager != null)
-            {
-                var channelId = notification.Android.ChannelId;
-                var channel = this.newManager.GetNotificationChannel(channelId);
-
-                if (channel == null)
-                {
-                    channel = new NotificationChannel(
-                        channelId,
-                        notification.Android.Channel,
-                        notification.Android.NotificationImportance.ToNative()
-                    );
-                    var d = notification.Android.ChannelDescription;
-                    if (!d.IsEmpty())
-                        channel.Description = d;
-
-                    this.newManager.CreateNotificationChannel(channel);
-                }
-                if (notification.Sound?.IsCustomSound() ?? false)
-                {
-                    var attributes = new AudioAttributes.Builder()
-                        .SetUsage(AudioUsageKind.NotificationRingtone)
-                        .Build();
-
-                    var uri = Android.Net.Uri.Parse(notification.Sound.CustomPath);
-                    channel.SetSound(uri, attributes);
-                    channel.EnableVibration(notification.Android.Vibrate);
-                    this.newManager.CreateNotificationChannel(channel);
-                }
-            
-                builder.SetChannelId(channelId);
-                this.newManager.Notify(notification.Id, builder.Build());
-            }
-            else
-            {
-                this.compatManager.Notify(notification.Id, builder.Build());
-            }
-#endif
         }
-
 
         protected virtual PendingIntent GetLaunchPendingIntent(Notification notification, string? actionId = null)
         {
@@ -289,14 +241,10 @@ namespace Shiny.Notifications
             PendingIntent pendingIntent;
             if ((notification.Android.LaunchActivityFlags & AndroidActivityFlags.ClearTask) != 0)
             {
-                pendingIntent = TaskStackBuilder
+                pendingIntent = AndroidX.Core.App.TaskStackBuilder
                     .Create(this.context.AppContext)
                     .AddNextIntent(launchIntent)
-#if ANDROIDX
                     .GetPendingIntent(notification.Id, (int)PendingIntentFlags.OneShot);
-#else
-                    .GetPendingIntent(notification.Id, PendingIntentFlags.OneShot);
-#endif
             }
             else
             {
@@ -317,11 +265,7 @@ namespace Shiny.Notifications
             if (colorResourceId <= 0)
                 throw new ArgumentException($"Color ResourceId for {colorResourceName} not found");
 
-#if ANDROIDX
-            return AndroidX.Core.Content.ContextCompat.GetColor(this.context.AppContext, colorResourceId);
-#else
-            return Android.Support.V4.Content.ContextCompat.GetColor(this.context.AppContext, colorResourceId);
-#endif
+            return ContextCompat.GetColor(this.context.AppContext, colorResourceId);
         }
 
 
@@ -454,7 +398,7 @@ namespace Shiny.Notifications
         protected virtual NotificationCompat.Action CreateTextReply(Notification notification, NotificationAction action)
         {
             var pendingIntent = this.CreateActionIntent(notification, action);
-            var input = new RemoteInput.Builder("Result")
+            var input = new AndroidX.Core.App.RemoteInput.Builder("Result")
                 .SetLabel(action.Title)
                 .Build();