+++ src/Shiny.Notifications/Platforms/Shared/NotificationJob.cs
@@ -22,7 +22,7 @@ namespace Shiny.Notifications
         }
 
 
-        public async Task<bool> Run(JobInfo jobInfo, CancellationToken cancelToken)
+        public async Task Run(JobInfo jobInfo, CancellationToken cancelToken)
         {
             var all = await this.repository.GetAll<Notification>();
             var pending = all
@@ -32,15 +32,12 @@ namespace Shiny.Notifications
                 )
                 .ToList();
 
-            var anyPending = false;
             foreach (var notification in pending)
             {
-                anyPending = true;
                 notification.ScheduleDate = null; // slight hack, kill schedule date as we're triggering now
                 await this.manager.Send(notification);
                 await this.repository.Remove<Notification>(notification.Id.ToString());
             }
-            return anyPending;
         }
     }
 }