+++ src/Shiny.TripTracker/TripTrackerLogCategory.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+
+namespace Shiny.TripTracker
+{
+    public static class TripTrackerLogCategory
+    {
+    }
+}